modify test c_dtype

+++ ops/built-in/tests/ut/ops_test/MatMulV2/test_MatMul_V2_impl.py
@@ -40,7 +40,7 @@ case4 = {"params": [{"shape": (64, 1280, 16, 16), "dtype": "float16", "format": 
                     {"shape": (64, 1280,16, 16), "dtype": "float16", "format": "FRACTAL_NZ", "ori_shape": (20480, 1024),"ori_format": "ND"},
                     None,
                     None,
-                    {"shape": (64, 64, 16, 16), "dtype": "float16", "format": "FRACTAL_NZ", "ori_shape": (1024, 1024),"ori_format": "ND"},
+                    {"shape": (64, 64, 16, 16), "dtype": "float32", "format": "FRACTAL_NZ", "ori_shape": (1024, 1024),"ori_format": "ND"},
                     True, False],
          "case_name": "MatMul_4",
          "expect": "success",
